+++ Form/ChoiceList/EnumChoiceList.php
@@ -10,6 +10,7 @@ use NetTeam\DDD\EnumUtil;
  * Choice list dla Enuma
  *
  * @author Paweł Macyszyn <pawel.macyszyn@netteam.pl>
+ * @author Dawid Drelichowski <dawid.drelichowski@netteam.pl>
  */
 class EnumChoiceList extends SimpleChoiceList
 {
@@ -35,15 +36,22 @@ class EnumChoiceList extends SimpleChoiceList
 
     private function translateArray(array $choices)
     {
+        $groups = false;
         $translatedChoices = array();
         foreach ($choices as $key => $choice) {
             if (is_array($choice)) {
+                $groups = true;
                 $translatedChoices[$key] = $this->translateArray($choice);
+                natsort($translatedChoices[$key]);
 
                 continue;
             }
 
             $translatedChoices[$key] = $this->translator->trans($this->prefix . $choice, array(), $this->domain);
+        }
+
+        if (!$groups) {
+            natsort($translatedChoices);
         }
 
         return $translatedChoices;