+++ Form/Type/EnumType.php
@@ -8,7 +8,8 @@ use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolverInterface;
 use Symfony\Component\OptionsResolver\Options;
 use Symfony\Component\Form\Exception\MissingOptionsException;
-use NetTeam\Bundle\DDDBundle\Form\DataTransformer\EnumToValueTransformer;
+use NetTeam\Bundle\DDDBundle\Form\DataTransformer\StringToEnumTransformer;
+use NetTeam\Bundle\DDDBundle\Form\DataTransformer\EnumsToValuesTransformer;
 use NetTeam\Bundle\DDDBundle\Form\ChoiceList\EnumChoiceList;
 
 class EnumType extends AbstractType
@@ -29,7 +30,13 @@ class EnumType extends AbstractType
             throw new MissingOptionsException('The option "class" must be defined', array('class'));
         }
 
-        $builder->addViewTransformer(new EnumToValueTransformer($options['class']), true);
+        if ($options['multiple']) {
+            $builder->addViewTransformer(new EnumsToValuesTransformer($options['class']), true);
+
+            return;
+        }
+
+        $builder->addViewTransformer(new StringToEnumTransformer($options['class']), true);
     }
 
     /**